requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
boto3==1.20.24
botocore==1.23.24
cffi==1.15.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==0.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.7
zope.interface==5.4.0
